+++ src/files_kraken/info.py
