+++ tests/test_data_organizer.py
@@ -3,12 +3,12 @@ import pathlib
 from dataclasses import dataclass, field
 from typing import ClassVar, List
 
-from src.blueprint import DataBlueprint
-from src.data_organizer import BlueprintsDBUpdater, BlueprintBuilder
-from src.database import DatabaseManager
-from src.fields import ParserField, DataParser
-from src.krakens_nest import Kraken
-from src.monitoring import Changes
+from src.files_kraken.blueprint import DataBlueprint
+from src.files_kraken.data_organizer import BlueprintsDBUpdater, BlueprintBuilder
+from src.files_kraken.database import DatabaseManager
+from src.files_kraken.fields import ParserField, DataParser
+from src.files_kraken.krakens_nest import Kraken
+from src.files_kraken.monitoring import Changes
 from test_database import db
