+++ tests/test_retools.py
@@ -1,4 +1,4 @@
-from src.retools import ReExecutor, SchemeMatcher, ReSorter, GroupSearcher
+from src.files_kraken.retools import ReExecutor, SchemeMatcher, ReSorter, GroupSearcher
 from test_collector import test_matcher
 
 # ReExecutor.fullmatch
